% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_profile_library.R
\name{af_profile_dir}
\alias{af_profile_dir}
\title{Get the global AF profile directory}
\usage{
af_profile_dir(create = TRUE)
}
\arguments{
\item{create}{Logical; create the directory if it does not exist.}
}
\value{
Invisibly returns the AF profile directory path.
}
\description{
Get the global AF profile directory
}
