% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{calculate_similarity_matrix}
\alias{calculate_similarity_matrix}
\title{Calculate Cosine Similarity Matrix}
\usage{
calculate_similarity_matrix(M)
}
\arguments{
\item{M}{Reference matrix (Markers x Detectors)}
}
\value{
A square matrix representing pairwise cosine similarities.
}
\description{
Calculate Cosine Similarity Matrix
}
