% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_matrix.R
\name{calculate_ssm}
\alias{calculate_ssm}
\title{Calculate Spectral Spread Matrix (SSM)}
\usage{
calculate_ssm(M, method = "OLS")
}
\arguments{
\item{M}{Reference matrix (Markers x Detectors)}

\item{method}{Unmixing method ("OLS" or "WLS")}
}
\value{
A matrix (Markers x Markers) representing unmixing spread
}
\description{
Quantifies the spreading error introduced by unmixing between fluorophores.
High values indicate that marker i significantly increases the noise in marker j.
}
\examples{
M <- matrix(c(1, 0.2, 0.1, 1), nrow = 2, byrow = TRUE)
rownames(M) <- c("FITC", "PE")
colnames(M) <- c("B2-A", "YG1-A")
ssm <- calculate_ssm(M)
ssm
}
