% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_profile_library.R
\name{delete_af_profile}
\alias{delete_af_profile}
\title{Delete a saved AF profile}
\usage{
delete_af_profile(name)
}
\arguments{
\item{name}{Profile name.}
}
\value{
Invisibly returns `TRUE` if the profile was deleted.
}
\description{
Delete a saved AF profile
}
