% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_profile_library.R
\name{load_af_profile}
\alias{load_af_profile}
\title{Load a saved AF profile}
\usage{
load_af_profile(name, show_plot = FALSE)
}
\arguments{
\item{name}{Profile name.}

\item{show_plot}{Logical; print the saved AF spectra plot after loading.}
}
\value{
A `spectreasy_af_profile` object.
}
\description{
Load a saved AF profile
}
