% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{plot_residuals_vs_expression}
\alias{plot_residuals_vs_expression}
\title{Plot Residuals vs Marker Expression}
\usage{
plot_residuals_vs_expression(
  results,
  marker,
  output_file = NULL,
  width = 160,
  height = 120
)
}
\arguments{
\item{results}{List of unmixed results}

\item{marker}{Name of marker to plot on X-axis}

\item{output_file}{Optional path to save the plot}

\item{width}{Plot width}

\item{height}{Plot height}
}
\value{
A `ggplot` object
}
\description{
Plot Residuals vs Marker Expression
}
