% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{plot_sample_rms_residuals}
\alias{plot_sample_rms_residuals}
\title{Plot Overall Detector Reconstruction Error Across Samples}
\usage{
plot_sample_rms_residuals(
  results,
  M = NULL,
  output_file = NULL,
  width = 225,
  height = 125
)
}
\arguments{
\item{results}{List of unmixed results from unmix_samples()}

\item{M}{Optional reference matrix to reconstruct raw intensities and compute relative error}

\item{output_file}{Optional path to save the plot. Set `NULL` to return the plot.}

\item{width}{Width of plot in mm}

\item{height}{Height of plot in mm}
}
\value{
A `ggplot` object.
}
\description{
Plot Overall Detector Reconstruction Error Across Samples
}
