% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{plot_similarity_matrix}
\alias{plot_similarity_matrix}
\title{Plot Cosine Similarity Matrix}
\usage{
plot_similarity_matrix(
  similarity_matrix,
  output_file = NULL,
  width = 180,
  height = 160
)
}
\arguments{
\item{similarity_matrix}{Matrix returned by calculate_similarity_matrix}

\item{output_file}{Optional path to save the plot. Set `NULL` to return the plot.}

\item{width}{Width of plot in mm}

\item{height}{Height of plot in mm}
}
\value{
A `ggplot` object.
}
\description{
Plot Cosine Similarity Matrix
}
