% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_matrix.R
\name{plot_ssm}
\alias{plot_ssm}
\title{Plot Spectral Spread Matrix}
\usage{
plot_ssm(SSM, output_file = NULL, width = 200, height = 180)
}
\arguments{
\item{SSM}{Matrix returned by calculate_ssm}

\item{output_file}{Optional path to save the plot. Set `NULL` to return the plot without writing a file.}

\item{width}{Width of plot in mm}

\item{height}{Height of plot in mm}
}
\value{
A `ggplot` object.
}
\description{
Plot Spectral Spread Matrix
}
\examples{
M <- matrix(c(1, 0.2, 0.1, 1), nrow = 2, byrow = TRUE)
rownames(M) <- c("FITC", "PE")
colnames(M) <- c("B2-A", "YG1-A")
ssm <- calculate_ssm(M)
p <- plot_ssm(ssm, output_file = NULL)
print(p)
}
