% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_analysis.R
\name{population_analysis_palettes}
\alias{population_analysis_palettes}
\title{Continuous palettes for population-analysis plots}
\usage{
population_analysis_palettes()
}
\value{
A named list of hexadecimal color vectors.
}
\description{
Continuous palettes for population-analysis plots
}
